demo/requirements.txt
Flask==2.2.2

opentelemetry-distro[otlp]
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging
